windows 7 taskbar fixes: http://i25.tinypic.com/egabsz.png

+++ ZScreenLib/Helpers/XMLSettings.cs
@@ -630,7 +630,7 @@ namespace ZScreenLib
                 }
             }
 
-            if (settingsFile != Program.appSettings.GetSettingsFilePath())
+            if (File.Exists(settingsFile) && settingsFile != Program.appSettings.GetSettingsFilePath())
             {
                 // Update AppSettings.xml
                 File.Copy(settingsFile, Program.appSettings.GetSettingsFilePath());
